moved parse props to own folder so it can be re used
