+++ exceptions/wql/parsing.py
@@ -1,0 +1,9 @@
+from exceptions.wiggle_base_exception import WiggleGraphBaseException
+
+
+class WiggleGraphIllegalPropertyValue(WiggleGraphBaseException):
+    pass
+
+
+class PropertyTypeAssignmentError(WiggleGraphBaseException):
+    pass
