+++ models/wql/parsing/tests/test_make_properties.py
@@ -1,7 +1,10 @@
 import pytest
 
-from exceptions.wql.make import MakePropertyTypeAssignmentError
-from models.wql.clauses.make.make_properties import ExtractedProperty, MakeProperty
+from exceptions.wql.parsing import PropertyTypeAssignmentError
+from models.wql.parsing.properties_pre import (
+    ExtractedPropertyPre,
+    WiggleGraphPropertyPre,
+)
 from models.wql.enums.property_type import PropertyType
 from testing.test_helpers import does_not_raise
 
@@ -11,32 +14,32 @@ class TestMakeProprietiesModels:
         "test_make_property, expected_extracted_property, exception",
         [
             pytest.param(
-                MakeProperty(
+                WiggleGraphPropertyPre(
                     property_name="foo", property_value="bar", string_type="bar"
                 ),
-                ExtractedProperty(
+                ExtractedPropertyPre(
                     property_value="bar", property_type=PropertyType.STRING_TYPE
                 ),
                 does_not_raise(),
                 id="EXP PASS: Correct assignment",
             ),
             pytest.param(
-                MakeProperty(
+                WiggleGraphPropertyPre(
                     property_name="foo",
                     property_value="bar",
                     string_type="bar",
                     bool_type="bar",
                 ),
                 None,
-                pytest.raises(MakePropertyTypeAssignmentError),
+                pytest.raises(PropertyTypeAssignmentError),
                 id="EXP EXEC: Two proprieties",
             ),
         ],
     )
     def test_emit_nodes(
         self,
-        test_make_property: MakeProperty,
-        expected_extracted_property: ExtractedProperty,
+        test_make_property: WiggleGraphPropertyPre,
+        expected_extracted_property: ExtractedPropertyPre,
         exception,
     ):
         with exception:
